* commit '7023df08f14ec5dee76ac54c03e870f84e297636':
  Revert "AArch64: Make graphics classes 64-bit compatible"

+++ core/java/android/view/GLES20RecordingCanvas.java
@@ -58,7 +58,7 @@ class GLES20RecordingCanvas extends GLES20Canvas {
         mDisplayList.clearReferences();
     }
 
-    long end(long nativeDisplayList) {
+    int end(int nativeDisplayList) {
         return getDisplayList(nativeDisplayList);
     }
