+++ graphics/java/android/graphics/LightingColorFilter.java
@@ -104,6 +104,6 @@ public class LightingColorFilter extends ColorFilter {
         nativeColorFilter = nCreateLightingFilter(native_instance, mMul, mAdd);
     }
 
-    private static native long native_CreateLightingFilter(int mul, int add);
-    private static native long nCreateLightingFilter(long nativeFilter, int mul, int add);
+    private static native int native_CreateLightingFilter(int mul, int add);
+    private static native int nCreateLightingFilter(int nativeFilter, int mul, int add);
 }
